backend/requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
pillow
numpy
pydantic
pydantic-settings
pandas
python-dotenv
cffi
cryptography
pytest
httpx
